% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwcorr.R
\name{cor.prob}
\alias{cor.prob}
\title{Calculate pairwise correlations}
\usage{
cor.prob(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A data.frame with columns h_var, v_var, and p.value
}
\description{
Internal function to calculate pairwise correlations and return p values
}
