% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder.R
\name{gladder}
\alias{gladder}
\title{Replica of Stata's gladder function}
\usage{
gladder(x)
}
\arguments{
\item{x}{A continuous numeric vector.}
}
\value{
A ggplot object with plots of each transformation
}
\description{
Creates ladder-of-powers histograms to visualize nine common transformations
and compare each to a normal distribution. The following transformations are
included: identity, cubic, square, square root, natural logarithm, inverse
square root, inverse, inverse square, and inverse cubic.
}
\examples{
gladder(iris$Sepal.Length)
gladder(mtcars$disp)

}
