% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpR.R
\docType{package}
\name{helpR}
\alias{helpR}
\alias{helpR-package}
\title{helpR: Miscellaneous helper and SAS/Stata duplicate functions}
\description{
The helpR package contains helper functions for tasks such as making
 print-friendly bivariate tables, sample size flow counts, and more. It
 also contains R approximations of some common, useful SAS/Stata functions.
}
\section{Helper functions}{

The helper functions
}

