% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder.R
\name{ladder}
\alias{ladder}
\title{Replica of Stata's ladder function}
\usage{
ladder(x)
}
\arguments{
\item{x}{A continuous numeric vector.}
}
\value{
A data.frame
}
\description{
Searches the ladder of powers histograms to find a transformation to make
\code{x} normally distributed. The Shapiro-Wilkes test is used to assess for
normality. The following transformations are included: identity, cubic,
square, square root, natural logarithm, inverse square root, inverse, 
inverse square, and inverse cubic.
}
\examples{
ladder(iris$Sepal.Length)
ladder(mtcars$disp)

}
