% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nagelkerke.R
\name{nagelkerke}
\alias{nagelkerke}
\title{Calculate Nagelkerke pseudo r-squared}
\usage{
nagelkerke(mod)
}
\arguments{
\item{mod}{A \code{glm} model object, usually from logistic regression.
The model must have been fit using the \code{data} option, in order to
extract the data from the model object.}
}
\value{
Numeric value of Nagelkerke r-squared for the model
}
\description{
Calculate Nagelkerke pseudo r-squared from a fitted model object.
}
