% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_means.R
\name{proc_means}
\alias{proc_means}
\title{Replica of SAS's PROC MEANS}
\usage{
proc_means(df, vars = NULL, by = NULL, n = T, mean = TRUE,
  sd = TRUE, min = TRUE, max = TRUE, median = FALSE, q1 = FALSE,
  q3 = FALSE, iqr = FALSE, nmiss = FALSE, nobs = FALSE)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{vars}{A character vector of numeric variables to generate descriptive 
statistics for. If the default (\code{NULL}), all variables are included,
except for any specified in \code{by}.}

\item{by}{Discrete variable. Separate statistics will be produced for 
each level. Default \code{NULL} provides statistics for all observations.}

\item{n}{Display number of rows with values. Default \code{TRUE}.}

\item{mean}{Display mean value. Default \code{TRUE}.}

\item{sd}{Display standard deviation. Default \code{TRUE}.}

\item{min}{Display minimum value. Default \code{TRUE}.}

\item{max}{Display maximum value. Default \code{TRUE}.}

\item{median}{Display median value. Default \code{FALSE}.}

\item{q1}{Display first quartile value. Default \code{FALSE}.}

\item{q3}{Display third quartile value. Default \code{FALSE}.}

\item{iqr}{Display interquartile range. Default \code{FALSE}.}

\item{nmiss}{Display number of missing values. Default \code{FALSE}.}

\item{nobs}{Display total number of rows. Default \code{FALSE}.}
}
\value{
A data.frame with columns variable, \code{by} variable, and 
  a column for each summary statistic.
}
\description{
Descriptive statistics for continuous variables, with the option of 
stratifying by a categorical variable.
}
\examples{
proc_means(iris, vars = c("Sepal.Length", "Sepal.Width"))
proc_means(iris, by = "Species")

}
