% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_flow.R
\name{sample_flow}
\alias{sample_flow}
\title{Create table illustrating sample exclusions}
\usage{
sample_flow(df, exclusions = c())
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{exclusions}{Character vector of logical conditions indicating which rows
should be excluded from the final sample. Exclusions occur in the order
specified.}
}
\value{
A data.frame with columns Exclusion, 'Sequential Excluded', and
  'Total Excluded' for display.
}
\description{
Generate a table illustrating sequential exclusion from an analytical sample
  due to user specified exclusions.
}
