% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_tidy.R
\name{stata_tidy}
\alias{stata_tidy}
\title{Tidy model output into similar format from Stata}
\usage{
stata_tidy(mod, var_label_df = NULL)
}
\arguments{
\item{mod}{A fitted model object}

\item{var_label_df}{A data.frame or tibble with columns "variable" and
"label" that contains display labels for each variable in \code{mod}.}
}
\value{
A data.frame with columns term and display
}
\description{
Create a display data frame similar to Stata model output for a fitted
  R model.
}
