% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_freq.R
\name{univar_freq}
\alias{univar_freq}
\title{Univariate statistics for a discrete variable}
\usage{
univar_freq(df, var, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{var}{A discrete, numeric variable.}

\item{na.rm}{logical. Should missing values (including \code{NaN}) be
removed?)}
}
\value{
A data.frame with columns \code{var}, NObs, and Percent
}
\description{
Descriptive statistics (N,%) for a continuous variable
}
\examples{
univar_freq(iris, var = "Species")
univar_freq(mtcars, var = "cyl")

}
