% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_compare.R
\name{bivariate_compare}
\alias{bivariate_compare}
\title{Create publication-style table across one categorical variable}
\usage{
bivariate_compare(df, compare, normal_vars = NULL,
  non_normal_vars = NULL, cat_vars = NULL, fisher = FALSE,
  workspace = NULL, var_label_df = NULL, include_na = FALSE)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{compare}{Discrete variable. Separate statistics will be produced for
each level, with statistical tests across levels.}

\item{normal_vars}{Normally distributed continuous variables that will be
included in the descriptive table.}

\item{non_normal_vars}{Non-normally distributed continuous variables that
will be included in the descriptive table.}

\item{cat_vars}{Categorical variables that will be included in the
descriptive table.}

\item{fisher}{Logical. Should Fisher's exact test be used for categorical
variables? Default \code{FALSE}.}

\item{workspace}{Numeric variable indicating the workspace to be used for
Fisher's exact test. If \code{NULL}, the default, the default value of
\code{2e5} is used.}

\item{var_label_df}{A data.frame or tibble with columns "variable" and
"label" that contains display labels for each variable specified in
\code{normal_vars}, \code{non_normal_vars}, and \code{cat_vars}.}

\item{include_na}{Logical. Should \code{NA} values be included in the
table and accompanying statistical tests? Default \code{FALSE}.}
}
\value{
A data.frame with columns label, overall, a column for each level
  of \code{compare}, and p.value. For \code{normal_vars}, mean (SD) is
  displayed, for \code{non_normal_vars} median (IQR) is displayed, and for
  \code{cat_vars} n (%) is displayed.
}
\description{
Descriptive statistics for categorical variables as well as normally and
non-normally distributed continiuous variables, split across levels of
a categorical variable. Depending on the variable type, an appropriate
statistical test is used to assess differences across levels of the
comparison variable.
}
\examples{
bivariate_compare(iris, compare = Species, normal_vars = c("Sepal.Length", "Sepal.Width"))

bivariate_compare(mtcars, compare = cyl, non_normal_vars = "mpg")
}
