% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_flow.R
\name{sample_flow}
\alias{sample_flow}
\title{Create table illustrating sample exclusions}
\usage{
sample_flow(df, na_vars)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{na_vars}{An ordered character vector containing variables that should
excluded due to having missing values (\code{NA}, \code{NaN}). Exclusions
occur in the order specified.}
}
\value{
A data.frame with columns variable, 'Sequential Missing', and 
  'Total Missing' for display.
}
\description{
Generate a table illustrating sequential exclusion from an analytical sample
  due to missing values.
}
